requirements.txt
# python3 -m  pipreqs.pipreqs . --force
gradio==3.24.1              # User interface
kornia==0.6.9               # Degradations on dataset
matplotlib==3.6.2           # Plotting loss
numpy==1.23.4               # General data manipulation
opencv_python==4.6.0.66     # Reading images, jpeg compression
Pillow==9.5.0               # Reading images
tifffile==2023.3.21         # Writing results with tiles to reduce memory usage
timm==0.6.12                # EfficientNet
torch==2.0.0          # Training and inference
torchvision==0.15.1   # Perceptual loss
imagecodecs==2023.3.16      # Reading images
